north_Assignment/requirements.txt
Django
python-decouple==3.8
gunicorn
psycopg2-binary==2.9.9
requests
djangorestframework
django-cors-headers
django-oauth-toolkit
django-rest-framework-social-oauth2
social-auth-app-django
redis
whitenoise
drf-spectacular==0.27.1
python-decouple==3.8
google-api-python-client
google-auth-oauthlib>=0.4.6
google-auth-httplib2>=0.1.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
